+++ app/models/design.rb
@@ -255,29 +255,20 @@ class Design < ActiveRecord::Base
     end
   end
 
+  def latex_partial(partial)
+    File.read(File.join('app', 'views', 'designs', 'latex', "_#{partial}.tex.erb"))
+  end
+
   def latex_file_location(current_user)
-    # @design = self
     jobname = "design_#{self.id}"
-    root_folder = FileUtils.pwd
-    output_folder = File.join(root_folder, 'tmp', 'files', 'tex')
-    template_folder = File.join(root_folder, 'app', 'views', 'designs')
-    file_template = File.join(template_folder, 'print.tex.erb')
-    file_tex = File.join(root_folder, 'tmp', 'files', 'tex', jobname + '.tex')
-    file_in = File.new(file_template, "r")
-    file_out = File.new(file_tex, "w")
-    template = ERB.new(file_in.sysread(File.size(file_in)))
-    file_out.syswrite(template.result(binding))
-    file_in.close()
-    file_out.close()
+    output_folder = File.join('tmp', 'files', 'tex')
+    file_tex = File.join('tmp', 'files', 'tex', jobname + '.tex')
 
-    # Run twice to allow LaTeX to compile correctly (page numbers, etc)
-    `#{LATEX_LOCATION} -interaction=nonstopmode --jobname=#{jobname} --output-directory=#{output_folder} #{file_tex}`
-    `#{LATEX_LOCATION} -interaction=nonstopmode --jobname=#{jobname} --output-directory=#{output_folder} #{file_tex}`
+    File.open(file_tex, 'w') do |file|
+      file.syswrite(ERB.new(latex_partial('print')).result(binding))
+    end
 
-    # Rails.logger.debug "----------------\n"
-    # Rails.logger.debug "#{LATEX_LOCATION} -interaction=nonstopmode --jobname=#{jobname} --output-directory=#{output_folder} #{file_tex}"
-
-    file_pdf_location = File.join('tmp', 'files', 'tex', "#{jobname}.pdf")
+    Design.generate_pdf(jobname, output_folder, file_tex)
   end
 
 
@@ -307,14 +298,7 @@ class Design < ActiveRecord::Base
     file_in.close()
     file_out.close()
 
-    # Run twice to allow LaTeX to compile correctly (page numbers, etc)
-    `#{LATEX_LOCATION} -interaction=nonstopmode --jobname=#{jobname} --output-directory=#{output_folder} #{file_tex}`
-    `#{LATEX_LOCATION} -interaction=nonstopmode --jobname=#{jobname} --output-directory=#{output_folder} #{file_tex}`
-
-    # Rails.logger.debug "----------------\n"
-    # Rails.logger.debug "#{LATEX_LOCATION} -interaction=nonstopmode --jobname=#{jobname} --output-directory=#{output_folder} #{file_tex}"
-
-    file_pdf_location = File.join('tmp', 'files', 'tex', "#{jobname}.pdf")
+    Design.generate_pdf(jobname, output_folder, file_tex)
   end
 
   def load_variables